pyproject.toml
[project]
name = "workshop-planner-agent"
version = "0.1.0"
description = "LangChain v1 workshop showcasing agent capabilities"
requires-python = ">=3.10"
dependencies = [
    "langchain>=1.0.0",
    "langchain-openai>=1.0.0",
    "langchain-anthropic>=1.0.0",
    "langgraph>=1.0.0",
    "langchain-mcp-adapters>=0.2.1",
    "python-dotenv>=1.0.0",
    "langsmith-fetch>=0.3.1",
    "ipykernel>=7.1.0",
    "jupyter>=1.1.1",
]

